Migrated network layout to class pattern.
